% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules-line-break.R
\name{n_lines}
\alias{n_lines}
\title{Count the number of line breaks in a parse table and all its children}
\usage{
n_lines(pd)
}
\arguments{
\item{pd}{A parse table}
}
\description{
This can be useful if styling depends on the number of lines within an
expression.
}
\keyword{internal}
