% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-export.R
\name{style_dir}
\alias{style_dir}
\title{Like \code{\link[styler:style_dir]{styler::style_dir()}}, but \code{style} defaulting to \code{semicolon_style}}
\usage{
style_dir(...)
}
\arguments{
\item{...}{Passed to \code{\link[styler:style_dir]{styler::style_dir()}}.}
}
\description{
Like \code{\link[styler:style_dir]{styler::style_dir()}}, but \code{style} defaulting to \code{semicolon_style}
}
