% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-export.R
\name{style_file}
\alias{style_file}
\title{Like \code{\link[styler:style_file]{styler::style_file()}}, but \code{style} defaulting to \code{semicolon_style}}
\usage{
style_file(...)
}
\arguments{
\item{...}{Passed to \code{\link[styler:style_file]{styler::style_file()}}.}
}
\description{
Like \code{\link[styler:style_file]{styler::style_file()}}, but \code{style} defaulting to \code{semicolon_style}
}
