% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-export.R
\name{style_pkg}
\alias{style_pkg}
\title{Like \code{\link[styler:style_pkg]{styler::style_pkg()}}, but \code{style} defaulting to \code{semicolon_style}}
\usage{
style_pkg(...)
}
\description{
Like \code{\link[styler:style_pkg]{styler::style_pkg()}}, but \code{style} defaulting to \code{semicolon_style}
}
